Updates to JSBSim and FDM interface.

+++ src/FDM/JSBSim/FGAtmosphere.h
@@ -141,19 +141,19 @@ public:
   inline float GetWindPsi(void) { return psiw; }
   
 private:
-  float rho;
+  double rho;
 
   int lastIndex;
-  float h;
-  float htab[8];
-  float SLtemperature,SLdensity,SLpressure,SLsoundspeed;
-  float rSLtemperature,rSLdensity,rSLpressure,rSLsoundspeed; //reciprocals
-  float temperature,density,pressure,soundspeed;
+  double h;
+  double htab[8];
+  double SLtemperature,SLdensity,SLpressure,SLsoundspeed;
+  double rSLtemperature,rSLdensity,rSLpressure,rSLsoundspeed; //reciprocals
+  double temperature,density,pressure,soundspeed;
   bool useExternal;
-  float exTemperature,exDensity,exPressure;
+  double exTemperature,exDensity,exPressure;
   
   FGColumnVector3 vWindNED;
-  float psiw;
+  double psiw;
 
   void Calculate(float altitude);
   void Debug(void);
